// src/types/models.ts
export interface ISelectedAgent {
    name: string;
    address: string;
    feeBIPS: string;
    handshakeType: number;
    underlyingAddress: string;
    infoUrl: string;
}
// ...
export interface IUtxo {
    address?: string;
    hexTx: string;
    index: number;
    txid: number;
    utxoAddress: string;
    value: string;
    vout: number;
}